requirements.txt
pandas
pyyaml
opencv-python
numpy
tqdm
scipy
pillow